libhorde: add 'transcript' option to record all incoming messages to a file
This is so that a module 'foo' can be debugged by adding, say,
 (transcript /var/log/horde/foo)
to foo.rc; this will then record the session, which can be replayed later
with 'foo </var/log/horde/foo' (possibly inside gdb or with debugging code
added).

Note that there is a possible security issue here: if it's possible for an
attacker to modify a module's .rc file, they can add a (transcript) pointing
to an arbitrary file which (if the module has write permissions there) will
be overwritten with hmsgs.